Valida as roles do utilizador para deletar ou alterar informacoes

+++ Microservices/Controllers/RegisterController.cs
@@ -1,9 +1,12 @@
 ﻿using AutoMapper;
+using MailKit;
 using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Mvc;
 using System.IdentityModel.Tokens.Jwt;
+using System.Net;
 using System.Security.Claims;
 using WebApiTemplate.Dtos;
+using WebApiTemplate.Models;
 using WebApiTemplate.Services;
 
 namespace WebApiTemplate.Controllers
@@ -18,18 +21,21 @@ namespace WebApiTemplate.Controllers
         private readonly IUserTransactionalService _userTransactionalService;
         private readonly IMapper _mapper;
         private readonly Helpers.JwtAuthenticationManager _jwtAuthenticationManager;
+        private readonly Helpers.AuthorizationHelper _authorizationHelper;
 
         public RegisterController(IMapper mapper,
             IRegisterService registerService,
             IFilterService filterService,
             IUserTransactionalService userTransactionalService,
-            Helpers.JwtAuthenticationManager jwtAuthenticationManager)
+            Helpers.JwtAuthenticationManager jwtAuthenticationManager,
+            Helpers.AuthorizationHelper authorizationHelper)
         {
             _registerService = registerService;
             _filterService = filterService;
             _userTransactionalService = userTransactionalService;
             _mapper = mapper;
             _jwtAuthenticationManager = jwtAuthenticationManager;
+            _authorizationHelper = authorizationHelper;
         }
 
         [HttpGet("GetAll")]
@@ -50,31 +56,23 @@ namespace WebApiTemplate.Controllers
 
             var getAll = _registerService.GetAll();
 
-            Request.Headers.TryGetValue("Authorization", out var headerValue);
+            User user = _authorizationHelper.GetAuthorization(Request.Headers);
 
-            var handler = new JwtSecurityTokenHandler();
-
-            var jwtTokenId = "";
-            try
+            if (user != null)
             {
-                var jwtToken = handler.ReadJwtToken(headerValue.ToString().Split(" ")[1]);
-                jwtTokenId = jwtToken.Claims.First(claim => claim.Type == "certserialnumber").Value;
-            }
-            catch (Exception e)
-            {
-                throw new ArgumentNullException(nameof(e));
-            }
-
-            if (!String.IsNullOrEmpty(jwtTokenId))
-            {
-                var userItem = _userTransactionalService.GetUserById(int.Parse(jwtTokenId));
-                if (userItem != null && userItem.RoleId == 2)
+                //var userItem = _userTransactionalService.GetUserById(int.Parse(jwtTokenId));
+                if (user != null && user.RoleId == 2)
                 {
-                    var userZone = _userTransactionalService.GetZoneByUser(int.Parse(jwtTokenId));
+                    //var userZone = _userTransactionalService.GetZoneByUser(int.Parse(jwtTokenId));
+                    var userZone = _userTransactionalService.GetZoneByUser(user.Id);
 
                     //Get Zone
                     getAll = getAll.Where(r => r.ZoneId == userZone.ZoneId); //TODO
                 }
+            }
+            else
+            {
+                throw new ArgumentNullException("Invalid Object");
             }
 
             //var jti = jwtToken.Claims.First(claim => claim.Type == "sid").Value;
@@ -84,6 +82,27 @@ namespace WebApiTemplate.Controllers
             return Ok(_mapper.Map<IEnumerable<RegisterSmallDto>>(getAll));
         }
 
+        //private string GetAuthorization(IHeaderDictionary headers)
+        //{
+        //    //Request.Headers.TryGetValue("Authorization", out var headerValue);
+        //    headers.TryGetValue("Authorization", out var headerValue);
+
+        //    var handler = new JwtSecurityTokenHandler();
+
+        //    var jwtTokenId = "";
+        //    try
+        //    {
+        //        var jwtToken = handler.ReadJwtToken(headerValue.ToString().Split(" ")[1]);
+        //        jwtTokenId = jwtToken.Claims.First(claim => claim.Type == "certserialnumber").Value;
+        //    }
+        //    catch (Exception e)
+        //    {
+        //        throw new ArgumentNullException(nameof(e));
+        //    }
+
+        //    return jwtTokenId;
+        //}
+
         [HttpGet("GetAllByFilter")]
         public ActionResult<IEnumerable<RegisterSmallDto>> GetAllByFilter(int filterId)
         {
@@ -92,31 +111,37 @@ namespace WebApiTemplate.Controllers
             //var getAll = _registerService.GetAll();
             var getAll = _registerService.GetAll().Where(r => getAllFiltersId.Contains(r.Id));
 
-            Request.Headers.TryGetValue("Authorization", out var headerValue);
+            //Request.Headers.TryGetValue("Authorization", out var headerValue);
 
-            var handler = new JwtSecurityTokenHandler();
+            //var handler = new JwtSecurityTokenHandler();
 
-            var jwtTokenId = "";
-            try
-            {
-                var jwtToken = handler.ReadJwtToken(headerValue.ToString().Split(" ")[1]);
-                jwtTokenId = jwtToken.Claims.First(claim => claim.Type == "certserialnumber").Value;
-            }
-            catch (Exception e)
-            {
-                throw new ArgumentNullException(nameof(e));
-            }
+            //var jwtTokenId = "";
+            //try
+            //{
+            //    var jwtToken = handler.ReadJwtToken(headerValue.ToString().Split(" ")[1]);
+            //    jwtTokenId = jwtToken.Claims.First(claim => claim.Type == "certserialnumber").Value;
+            //}
+            //catch (Exception e)
+            //{
+            //    throw new ArgumentNullException(nameof(e));
+            //}
 
-            if (!String.IsNullOrEmpty(jwtTokenId))
+            User user = _authorizationHelper.GetAuthorization(Request.Headers);
+
+            if (user != null)
             {
-                var userItem = _userTransactionalService.GetUserById(int.Parse(jwtTokenId));
-                if (userItem != null && userItem.RoleId == 2)
+                //var userItem = _userTransactionalService.GetUserById(int.Parse(jwtTokenId));
+                if (user != null && user.RoleId == 2)
                 {
-                    var userZone = _userTransactionalService.GetZoneByUser(int.Parse(jwtTokenId));
+                    var userZone = _userTransactionalService.GetZoneByUser(user.Id);
 
                     //Get Zone
                     getAll = getAll.Where(r => r.ZoneId == userZone.ZoneId); //TODO
                 }
+            }
+            else
+            {
+                throw new ArgumentNullException("Invalid Object");
             }
 
             //getAll = getAll.Where(r => getAllFilters.Contains(r.Id));
@@ -150,17 +175,36 @@ namespace WebApiTemplate.Controllers
         [HttpDelete]
         public async Task<ActionResult> Delete(int id)
         {
-            _registerService.Delete(id);
+            User user = _authorizationHelper.GetAuthorization(Request.Headers);
 
-            return Ok();
+            if (user != null && user.RoleId == 1)
+            {
+                _registerService.Delete(id);
+
+                return Ok();
+            }
+            else
+            {
+                return BadRequest(new UnauthorizedAccessException());
+            }
         }
 
         [HttpPut]
         public async Task<ActionResult> Update(int id, [FromBody] RegisterIn entityIn)
         {
-            _registerService.Update(id, entityIn);
+            User user = _authorizationHelper.GetAuthorization(Request.Headers);
 
-            return Ok();
+            if (user != null && user.RoleId == 1)
+            {
+                _registerService.Update(id, entityIn);
+
+                return Ok();
+            }
+            else
+            {
+                //return Unauthorized("Not Authorized");
+                return BadRequest(new UnauthorizedAccessException());
+            }
         }
     }
 }
